3A102b: Minor tweak on line usage.
Now a grey line will appear between the header and the content, if
there is any content. A black line will always appear at the end of any
post. This will give the user an indication if there is any media to
expect, as existance of medias will cause the distance between the
lines increase drastically.